[Tools] Add a BBQ stress test
This is the initial version of a stress test for the BarbequeRTRM which spawn
a random number of BbqueRTLibTestApp applications. This number could be
fixed by specifying a value for the APPS environment variable, e.g.:
APPS=2 ./out/sbin/bbque-stest
will run a stress test with 2 concurrent applications.

In order to test different possible run-time scenarios, each application is
started and with a randomized workload and cycles-count demand.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>