[BospShell] Initial version
This add an initial version of a possible BOSP-Shell. The goal is to provide
a customized shell experience to the BOSP user where a set of useful commands
and operations are on hand by means of properly defined alias and predefined
functions.

To access the features offered by this shell, the user has just to source the
configuration file which is deployed with the Barbeque installation, i.e.:
$> . <BOSP_BASEDIR>/etc/bbque/bosp_init.env

Signed-off-by: Patrick Bellasi <derkling@gmail.com>