[MCT_Congestion] Optimize conversion of MSL parameters
The parameters regarding the Maximum Saturation Level of each resource type
are parsed by YaMS Scheduling Policy, from the configuration file, and then
passed to the instance of the metrics contribute. These are uint16_t values
that must be internally handled as percentages, that is numbers in range
[0,1].

To avoid division by 100 each time the parameters must be used,
MetricsContribute class now stores an array of float instead of uint16_t.
Thus, when the array of parameters in passed (in the constructor) they
are converted and stored in the new array.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>