[MCT] Change one GetUsageRegion() argument
The method now gets a reference to an EvalEntity_t instead of a pointer to
ApplicationConfIF. This because the query on the resource availability is
no more performed without considering the resource amount previously
allocated to the application to evaluate.

The reference to the EvalEntity_t is exploited to print out a more clean
and uniformed debug message, thanks to its method StrID().

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>