[MCT_Value] Goal gap support
The metrics contribute computation now takes into account the setting of a
"Goal gap". In this case a "target resource usage" is computed for each
resource, starting from the usage amounts of the current AWM and
multiplying it for the goal gap percentage.

If all the related resource usage of the AWM under evaluation are greater
or equal to the "target resource usages", the AWM static value is returned
as contribute index. Otherwise the index is computed starting from the
static value, applying a sort of "penalization". The metrics contribute
will be the most penalized index computed among all the resource types.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>