[MCT_Fairness] Fairness metrics contribute
Introduced the "fairness" contribute.
The expected behavior is to return higher contribute indexes for
scheduling choices that assign a "fair" amount of resource to the
application/EXC, taking into account the amount of applications to schedule
(in the same priority class). Choices of assignment that overpass an
identified quota (i.e. fair partition) are penalized exponentially.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>