[RM] Integrate ProfileManager core module
This ProfileManager is a core module in charge to collect statistics and
figure of merits mainly related to scheduling activities.  It provides a
method to compute scheduler performance indexes which must be called
after each scheduler activation. However, since a scheduling decision
could require a synchronization, in order to reduce latencies, the
scheduler performances evalutation are performed just after the pending
synchronization has been completed.

TODO: the scheduler performance indexes could be eventually computed
before a synchronization to conditionally activate it based on an the
system-wide optimization goals defined by an optimization policy.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>