[APP] Goal gap reset
The goal gap is reset once the scheduling is committed and thus
the applications are synchronized. However the reset is performed only if
the new AWM assigned (next AWM) shows a "value" higher than the current one.
This because we assumed that only in this case makes sense to silence the
better service request.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>