[Recipe] Change AWM value normalization
The normalization of the AWM values is now performed by dividing the value
parsed from the recipe for the maximum of the values. This leads to a
linear distribution of the values, which the previous approach did not
guarantee. Conversely, the new approach does not behave in a robust manner
in cases of recipes with AWMs having very close values, or even the same.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>