[Metrics 1/5] Add indexed metrics support
This patch add each metrics class the possibility to collect a set of
"indexed" sub-metrics. Once a metrics is created we have the possibility
to define a certain number (sm_count) of sub-metrics with a
corresponding set of description strings (sm_descs).

By default sub-metrics are not required, in this case the metrics behave
as before. Instead, if a set of sub-metrics is required, once the
metrics is required an vector of class-specific sub-metrics is created.
In this case, the addition of a new value requires the specification of
an index for the specific sub-metrics to be updated, and the same value
will be added both to the specified sub-metrics and the base metrics.
This allows to collect specialized values and also an aggregate value on
the base metric.

Already defined metrics have been converted to classes, which allows for
a cleaner implementation of the sub-metrics support.
Metrics constructors are moved into source file for a better
encapsulation of the implementation. This is the last patch which
provides unaltered features with respect to the original metrics
implementation without sub-metrics support.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>