[Metrics 3/5] Add a set of sub-metrics dumpers
This provides a set of sub-metrics dumping functions which provides a
proper formatting for both the name and description of a sub-metric as
well as the computation of the corresponding statistics.  The computed
statistics are dumped on the logfile.

TODO: perhaps a better usage of tempaltes should allows for a reduction
of code duplications on these fragments.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>