[Tools_STest] Add execution time control
This defines a new console tunable variable to define an upper bound for the
minimum execution time. Indeed, the actual execution time of an instance of
the test application is defined as a random value with this parameter as
upper bound.

This is particularly useful to run very short or very long tests.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>