[Configure] Add RPC Timeout configuration
The RPC channel could be subject to timeouts, e.g. while waiting for a
response to a command, which should be properly handled since impact directly
on framework responsiveness.
This patch defines a new configuration option to define the timeouts to use
in [ms] for each RPC communication. A conservative 1[s] timeout should be
reasonable safe also for highly congestion ed communications period.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>