[Deferrable] Add deferrable execution base class
A deferrable is an object which allows to execute a specified method at the
nearest deadline among a collection of schedule request.
The deferrable could be created or configured to provide a "periodic" execution
of a certain function. Schedule requests are collected and aggregated in order
to schedule the nearest request (in time) among all those asserted.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>