[RM] Switch to deferrable optimizations
This patch introduces the usage of deferrable optimizations. The call to
the Optimization method is now managed by a deferrable object which provides
the required support to collect and aggregate multiple scheduling events
and process them according to certain timing policy.

The timing policy is not yet provided, right now we use just the "on-demand"
mode and optimizations are scheduled as previously.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>