[RPC_FIFO] Ignore SIGPIPE
When BBQ writes to a pipe that has been closed by the remote application
(e.g. due to a crash) we receive this signal and BBQ is terminated.
For simple programs, this is often an appropriate default action.

For a multi-threaded program, the correct action is usually to ignore the
SIGPIPE signal, so that writing to a closed socket/fifo will not terminate
the program. The call will return EPIPE error, which could be actually
managed within the program to take the proper countermeasures.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>