[RPC] Report exit codes
The two methods, defined by the RPC channel interface, to send and receive
messages have been updated to return exit codes on errors.
As before, a positive number represent the count of bytes send/received.
While, if we get a negative number, that's an error condition.

For example this is used by the RPC FIFO plugin to return EPIPE error on a
broken pipe.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>