[AM+APP] Bugfix multiple removal commands
A fault could happen if, while an EXC is being removed, another remove command
is executed. This could happens for example on heavy congested scenarios,
running on a many-core machine, when a burst of applications exit the system
at the same time. In this case, due to RTLIB timeouts the unregister command
could be issued multiple time thus producing an overlapped execution within
BBQ.

This patch exploit the serialized access to the Application::Terminate method
to set the application state and check if this state is already FINISHED
to avoid multiple executions of ApplicationManager::DestroyEXC.
A warning is issued on multiple destroy.

TODO: fix the register/unregister critical paths by making them atomics and
with really low latencyes to avoid RTLIB timeouts.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>