[MCT_Congestion] Fix exp function parameters setting
This fixes a mistake in the setting of parameters for the exponential
filter function. This was due to a pair of subtractions between unsigned
integers, where the minus term is always bigger than the minuend. The
result was erroneously stored in an unsigned integer variable, producing a
wrong implicit cast.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>