[PP] Move PSD initialization out of critical path
The creation of an EXC is a critical path. Indeed, to register successfully
a burst of starting applications, without incurring into an RTLib TIMEOUT,
the ApplicationManager::CreateEXC must complete as soon as possible.
Setting-up Platform Specific Data (PSD) into this method could introduce
significant delays. This happens for instance using the x86 PIL, which
requires ~10ms per CGroup creation.

This patch moves the setup of PSD into the PlatformProxy; the first time
an application is mapped on the platform the PDS are initialized.
The main advantages of such an approach are:
1. does not impact the registration critical path, thus allowing for a safer
registration on burst with an higher number of applications;
2. reduce the number of re-schedule since it allows to enter more
applications before a schedule is triggered;
3. the time required to setup PSD is actually spread, since on burst its more
likely that only some applications will actually enter the system;
4. the time required to setup PSD does not impact anymore on RTLib TIMEOUTS
since resource mapping for a starting happens does not have such constraints;

Signed-off-by: Patrick Bellasi <derkling@gmail.com>