[AM] Add EXCs Cleanup deferrable
A defarrable Cleanup routine has been added to reduce the critical path
related to the DestroyEXC calls. The deferrable is used to asynchronously
perform all the operations, related to releasing applications resources, which
could be postponed in order to have a fast response to an RTLib command.

The Cleanup deferrable will be in charge for asynchronous release of Platform
specific data and removal of the application from the BBQ internal data
structures.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>