[AM] Switch to deferrable EXCs cleanup
In order to reduce the RTLib response critical path, now the DestroyEXC
call is used to mark the application as FINISHED and to schedule a
deferrable Cleanup.

All the time consuming operations, e.g. release of Platform Specific data,
will now be performed in deferred time by the Cleanup routine.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>