[SchedPol_YaMS] Bugfix ordering function
The ordering of the scheduling entities where not correctly performed in
the case of equal metrics. In this case indeed, the expected behavior is
that if one of the two scheduling entities has a NAP assertion, it must be
favored. Otherwise, if no NAPs are in place, the scheduling entity with
the greater AWM static value is ordered first.

This behavior was not correctly performed, since in the first condition
compare the "=" case was included.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>