[RM] Bugfix segfault on event processing
On START event processing, if all the applications have exited at the time
of event processing, the am.HighestPrio call return a NULL reference.
This was not properly handled, now we ignore such events thus avoiding the
scheduling of a new optimization.

TODO: add a counter monitor to keep track of the occurence of such kind of
evente.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>