[RTLIB] Bugfix cycles count on AWM switch
When an AWM switch happens, the GetWorkingMode call returns a the request
of enter in SYNC_MODE but we still need to correctly account for
a previously completed processing cycle.
Thus, a call to  SyncTimeEstimation is issued to allows a properly collection
of related statistics. This should allows also to have a correct reporting
at the end of the application execution.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>