[RU] Class attributes made private
ResourceUsage has been modified in a previous release, being transformed
from a data structure to a class. Since some development branches used
it as a data structure, their attributes have been kept public, in order
to made the code still compilable. Right now, the methods properly
provided by the class are used in place of the direct access to the
attributes. Therefore the attributes can be moved under a 'private'
section.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>