[RTLIB] Add (pre) Release notification
This patch introduces a new onRelease() method to the BbqueEXC base class.
Such a method is intended to host all the clean-up and resources release code
of a BBQ application and it serves to two main purposes:
1. it provides a cleaner interface to the appplication developer, which is
based just on a call-back methods. Indeed, the developer is now encouraged to
put all application specific code into the provided based class methods, thus
avoiding the usage of the class destructor.
2. it provides an hook for the RTLib to properly release resources after the
user-code has completed and immediately before the application is going to
exit, e.g. collect execution metrics such as overall memory consumption.

The plain RTLib API introduces a new Release notifier callback which is intended
to be called in late termination stages, right before the application release
all its resources. For example, this is a suitable place-holder for memory
reporting code which is expected to be merged in a following patch.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>