[AEM_CPP] Bugfix race-condition on application unregister
The WaitCompletion is a blocking call that the main could use to know when
an application has completed the execution. After this code returns the
application could eventually call an explicit un-regsiter.
However, at the exit of the control loop the base-class code already
takes care of EXC unregistration and this produces a race-condition on
DEBUG compiled codes.

This patch solve this problem by adding a dedicated flag to notify the
completion of the workload processing.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>