[RTLIB] [MOST 1/3] Add MOST option flag
This patch extends the BBQUE_RTLIB_OPTS environment variable with the
possibility to request the MOST output syntax for metrics dumping at
application exit time.

A new "M" flags, with an optional tag string is now supported.
The tag is considered all the string following the "M" up to the next
":" or end of line.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>