[RecipeLoader_XML] Improve error handling
The error handling has been improved by adding an "error" block at the end
of the LoadRecipe() method. Therefore, whenever an error condition is
checked, the code jump to such block instead of returning immediately.
This is order to have a unique block of code where to perform cleaning
action before aborting the recipe loading.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>