[CPP11] Add CPP11 fallback support
Some platforms, notably Android, does not yet fully support some
features of the C++11 standard. This patch provides a fallback
implementation of all the features required for a successfull
BarbequeRTRM compilation on these platforms.

This fallback implementation is based on code from the GNU Standard C++
library, which has been adapted to relay on a plain C API, e.g. using
POSIX threads.

A set of convenient "dispatching headers" is also provided, which select
either the system library or the fallback implementation based on the
capabilities of the target platform. For the time being, only targets
defining __BIONIC__ make use of the fallback implementation.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>