[RTLIB_AS-RTM] Update getCurrentOP implementation
This commit slightly changes the behaviour of the function getCurrentOP.
The reason of this commit is that it is possible that the metrics giving
the order of the Operating Points doesn't follow the direction for the goals
declared for the application.
For this reason it is better to check both the lower and upper operating
points that satisfy the isValid condition. The choice of first checking
the lower OPs and then the highest OPs is based on a tradeoff between the
performance and the implementation time available for this function.
So far, the complexity is linear since in the worst case the function will
have to scan all the vector of OP but better implementation might be
possible.

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Acked-by: Andrea Di Gesare <andrea.digesare@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>