[RTLIB_AS-RTM] Add getNextOP function
This commit introduces the getNextOP that tries to make transparent to the
user the way in which the AS-RTM gets a new Operating Point.
So far this function uses getCurrentOP function and hence shares the same
potential of being improved allowing to have a shorter access time,
especially for the worst-case scenario.

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Acked-by: Andrea Di Gesare <andrea.digesare@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>