[RTLIB_AS-RTM 1/6] Add UPDATE_[UPPER|LOWER]_BOUND
This commit introduces two macros in the code of the AS-RTM that are used
to update the constraints used to select OperatingPoints.
As stated as well in the comments, a default action is taken when no
constraint of the given type (upper or lower) is found: a new constraint
with a strict inequality. It is just a default implementation: nothing avoids
to change it and switch to a normal "inequality" (>= vs > and <= vs <)

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Acked-by: Andrea Di Gesare <andrea.digesare@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>