[RTLIB_AS-RTM 3/6] Updates goal-checking functions
With the introduction of the fullCheckGoal function, a decision has been made
regarding the functions that might be called by the developer most of the
times. The decision of switching to relativeError as return value for most
of the function has been made since this is the value more understandable
for an human-eye. Anyway, thanks to the fullCheckGoal function, all the
information (such as NAP) will be available when needed.

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Acked-by: Andrea Di Gesare <andrea.digesare@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>