[RTLIB_AS-RTM 5/6] Implementation of the core features of the AS-RTM
This commit introduces all the new methods of the ApplicationRTM that
can be used to tune-up an application with the least effort of the developer.
With the use of these new methods, calling the checkGoal function and
then changing manually application parameters according to the status of
the goal achievement of the application is no longer necessary: the new
getNextOP will do everything in an automatic manner.

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Acked-by: Andrea Di Gesare <andrea.digesare@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>