[RTLIB_CPS 1/4] Add CPS enforcing routines
A couple of methods have been added which allows to configure an expected
Cycles Per Second value. When such a value is configured, at run-time
the RTLib provide support to monitor the actual cycles time and eventually
insert a properly computed delay in order to get the specified cycle
rate.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>