[Utility] Add BBQ configurable format string
This log format string is suitable to be used by BarbequeRTRM core modules.
For the time being, the string has been simplified to not dump the timestamp
since this is a costly operation and moreover, this format string is used
for a limited number of calls, i.e. when a logger module is not available.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>