[RTLIB_RPC] Bugfix statistics update
This patch fixes a race condition which could lead to statistics being
accessed before they are initialized. This happens for instance when multiple
applications starts concurrently, each one enabling an EXC but just one
sending a GWM while the others receive a PRECHANGE notify before sending
their GWM. This case could be noticed when the time required by the scheduler
triggered by the first schedule requires is faster then the time for the
other applications to send a GWM.

Statistics are initialized only when a valid AWM has been assigned.
This patch checks if a valid statistics are present before actually calling
a SyncTime estimation: if statistics are not initialized there is nothing to
be estimated.

This should fix a SEGFAULT which happens when the SyncTime estimation is called
for an application which:
1. is enabled
2. has not yet sent a GWM
3. has already received a PRECHANGE, thus entering sync mode

In these situations: statistics are not yet initialized and the application
is just starting, thus it is ok to return without any synctime estimation.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>