[RTLIB_RPC] Send GWM only if not yet in SYNC mode
It could happen that an enabled AWM gets a PRECHANGE, thus entering SYNC mode,
even before sending a GWM. This could happen for instance when multiple EXC
are enabled concurrently and a scheduled, triggered by one of them, completes
even before the others could request the schedule.

In these cases, i.e. already in SYNC mode, we could avoid to send a new GWM
request. Indeed the EXC is going to be already served.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>