[RecipeLoader_XML] No platform validation for TPD
The platform ID check is performed only if the framework has not been
compiled with the Test Platform Data configuration enabled. This patch
introduces such a behavior, by skipping the platform ID control in case of
TPD enabled, and proceeding with the first "platform" section declared in
the recipe.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>

Conflicts:

	plugins/rloader/xml_rloader.cc