[AWM] AddResourceUsage refactoring
The method does not check anymore if the amount of resource requirement
specified can be satisfied by the current hardware resources configuration.
Indeed, from now on we aim to load every working mode at first glance, and
then performing a validation in a next step.

However the existence check of the resource is still performed, but
exploiting the ExistResource() method instead of looking at the total
amount of resource available.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>