[RPC] Update messages priority
The RPC messages priority has been re-shuffled based on these principles:
1. higher ID messages have higer priority.
2. BBQ originated messages have higher priority than APPs originated.
3. SyncProtocol messages have a priority which is the inverse of the
corresponding protocol step; this allows to give precedence to late
responding apps in early steps of the sync protocol.
4. Apps "termination" messages have priority on apps "initialization";
thus allowing to eventually release resources before a schedule is
issued.
5. Apps "configuration" messages have priority on apps "scheduling";
thus allowsing to eventually tune apps requirements before a schedule is
trigghered.

Based on this schema: "scheduling" and "pairing" are the two lowest
priority messages. This should allows for a better handling of bust
of applications starting at the same time, by reducing and aggregating
as much as possible the scheduler activation.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>