[AWM] Introduce validation and hidden status
To support run-time changes in the total resources availability, the
concepts of "validation" and "hidden status" have been introduced.

The validation is performed by calling the method Validate(), and consists
of checking if all the resource requirements included in the AWM can be
satisfied, considering the current total amounts of hardware resources.
In case of unsatisfiable requirement the AWM is marked "hidden" by setting
the homonym boolean attribute to true. The method Hidden() allows to check
if the AWM is hidden or not.

Note: Obviously hidden AWM should not be taken into account by the scheduling
policy.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>