[App] Add method for AWMs reload
This is required whenever there has been a change in the status of the
resources (e.g., total availability) due to an HW fault or some low-level
actions of power/thermal management, once a re-validation of
the recipes has been performed. This last feature is going to be introduced
by next patches.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>