[Recipe] Add validation
The validation consists of two operations:
   - Validate each AWM, given the current resources availability
   - Normalize the AWM values

This is performed by the brand new method Validate(), which is currently
expected to be called by the Application Manager once the recipe has been
loaded.

Moreover the AWMs values normalization has been rearranged in order to
operate just on the set of not hidden AWMs. However the method
WorkingModesAll() returns the whole vector of AWMs. It is up to the
Application to consider only the not hidden AWMs for the scheduling.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>