[AP+YM] Merge Sync and Async calls
The asynchronous and synchronous version of synchronization protocol calls
are quite similar. This patch simplify the AP API by removing the "*_Async"
version of the calls and exploiting conditional compilation to select the
proper version at compile time.
This means that it is not more possible to dynamically switch at run-time
from sync and async version, but this is not considered a limitation so far.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>