[RTLIB_AS-RTM]  Bugfix vectors management
This patches a tricky problem probably related to a misused vector instead
of a simples linked list. Indeed, the client code of these vectors is using
push_back calls to put new elements into the vector.

Since elements in the vector are added via a push_back function, the
result of:
   vector.resize(1)
   vector.push_back(element)
will lead to a vector with two elements instead of one as wanted.

If random access is not required on these data structures we should better
switch to a list based implementation. However, for the time being, this
patch should solve some bugs.

More details are tracked in the bosp-devel ML:
https://groups.google.com/d/msg/bosp-devel/7K-tIuC7E10/UmFaK1vRoXUJ

Signed-off-by: Patrick Bellasi <derkling@gmail.com>