[RM] Add support for periodic optimizations
The scheduling of resources to applications could be configured to be
periodically triggered once every "opt_interval". This interval could
be defined in [ms] by a proper configuration file option.
A value of 0[ms], which is the default, could be used to disable the periodic
optimization thus using just an event based activation.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>