[RA] Add Count() member function
The function returns the number of resource descriptors referenced by a
given resource path. Usually, if the path is a template, the member
function should return a value greater than 1. Conversely, a specific path
should return a counting value equal to 1.

The function is available also to plugins, thus it has been exported into
the System class as member function ResourceCount().

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>